pyproject.toml
[project]
name = "comfyui_ssstereoscope_bysamseen"
description = "Create immersive 3D stereoscopic images and videos! Transform your ComfyUI generations into stunning side-by-side 3D visuals for videos and image sequences. Powered by Depth-Anything-V2, no external depth maps needed. Perfect for VR, 3D displays, and cross-eyed viewing - no special glasses required!"
version = "1.1.0"
license = "LICENSE"
dependencies = ["numpy", "Pillow", "torch", "transformers>=4.51.0", "opencv-python", "tqdm"]
keywords = ["stereoscopic", "3D", "SBS", "side-by-side", "depth-map", "VR", "virtual-reality", "ComfyUI", "depth-estimation", "3D-conversion", "depth-anything", "3D-video", "3D-image", "cross-eyed", "parallel-view", "AI-depth", "immersive", "no-glasses-3D"]

[project.urls]
Repository = "https://github.com/MrSamSeen/ComfyUI_SSStereoscope"
Documentation = "https://github.com/MrSamSeen/ComfyUI_SSStereoscope#readme"
Issues = "https://github.com/MrSamSeen/ComfyUI_SSStereoscope/issues"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "samseen"
DisplayName = "SideBySide_Stereoscope"
Icon = "sbs_icon.png"
Tags = ["3D", "stereoscopic", "depth-map", "VR", "video", "image-sequence", "depth-anything"]
